# Reduce allocations in copyTrailer (saves ~43% time)