```bash
 benchstat old.txt new.txt 
goos: darwin
goarch: arm64
pkg: github.com/valyala/fasthttp
cpu: Apple M3 Max
               │   old.txt   │               new.txt               │
               │   sec/op    │   sec/op     vs base                │
CopyTrailer-16   260.6n ± 1%   147.5n ± 0%  -43.42% (p=0.000 n=20)

               │   old.txt    │               new.txt               │
               │     B/op     │     B/op      vs base               │
CopyTrailer-16   1.109Ki ± 0%   1.125Ki ± 0%  +1.41% (p=0.000 n=20)

               │   old.txt   │              new.txt               │
               │  allocs/op  │ allocs/op   vs base                │
CopyTrailer-16   16.000 ± 0%   1.000 ± 0%  -93.75% (p=0.000 n=20)
```