@erikdubbelboer 2025-10-06
Sorry I meant it can better like this: https://github.com/valyala/fasthttp/commit/19624506292b2f05ed08774d1f6c814340efb409
This way it reuses the memory instead of making a new slice every time.